Get rid of test UI, it doesn't work
