--- UIDev/Framework/IPluginUIMock.cs
@@ -1,10 +1,0 @@
-﻿using ImGuiScene;
-using System;
-
-namespace UIDev
-{
-    interface IPluginUIMock : IDisposable
-    {
-        void Initialize(SimpleImGuiScene scene);
-    }
-}
